[#28] chore: mockdata 분리
